Remove CareMember service.
git-tfs-id: [http://hillstfs2013:8080/tfs/Phytel]$/PhytelCode/Phytel.Net/Services;C22457

+++ API/AppDomain/Phytel.API.AppDomain.NG/Development/Current/NG.DTOs/Contact/CareTeam/Member.cs
@@ -13,7 +13,7 @@ namespace Phytel.API.AppDomain.NG.DTO
         public bool Core { get; set; }
         public string Notes { get; set; }
         public string FrequencyId { get; set; }
-        public int? Distance { get; set; }
+        public double? Distance { get; set; }
         public string DistanceUnit { get; set; }
         public string ExternalRecordId { get; set; }
         public string DataSource { get; set; }
