+++ API/DataDomain/Phytel.API.DataDomain.Contact/Development/Current/Contact.DTO/CareTeam/CareTeamMemberData.cs
@@ -13,7 +13,7 @@ namespace Phytel.API.DataDomain.Contact.DTO.CareTeam
         public bool Core { get; set; }
         public string Notes { get; set; }
         public string FrequencyId { get; set; }
-        public int? Distance { get; set; }
+        public double? Distance { get; set; }
         public string DistanceUnit { get; set; }
         public string ExternalRecordId { get; set; }
         public string DataSource { get; set; }
